+++ src/components/ViewManager/views/ProtocolSummary.js
@@ -1,23 +1,6 @@
 import React, { useEffect, useState } from 'react';
-import { remote, ipcRenderer } from 'electron';
-import path from 'path';
-import fse from 'fs-extra';
-import os from 'os';
+import { ipcRenderer } from 'electron';
 import ProtocolSummary from '@app/lib/ProtocolSummary/ProtocolSummary';
-
-const print = () => {
-  const win = remote.BrowserWindow.getFocusedWindow();
-
-  win.webContents.printToPDF({ printBackground: true }).then((data) => {
-    const pdfPath = path.join(os.homedir(), 'Desktop', 'temp.pdf');
-    fse.writeFile(pdfPath, data, (error) => {
-      if (error) { throw error; }
-      console.log(`Wrote PDF successfully to ${pdfPath}`);
-    });
-  }).catch((error) => {
-    console.log(`Failed to write PDF to ${pdfPath}: `, error);
-  });
-};
 
 const ProtocolSummaryView = () => {
   const [protocol, setProtocol] = useState(null);
@@ -38,7 +21,6 @@ const ProtocolSummaryView = () => {
 
   return (
     <div>
-      <button type="button" onClick={() => print()}>Print</button>
       <ProtocolSummary protocol={protocol} />
     </div>
   );
