+++ src/lib/ProtocolSummary/styles/_protocol-summary-stage.scss
@@ -1,20 +1,32 @@
 .protocol-summary-stage {
-  margin-top: 8rem;
+  margin-bottom: 8rem;
+  page-break-before: always;
 
   &__heading {
-    background: pink;
+    background: var(--color-platinum--dark);
+    border-radius: 1rem;
     padding: 1rem 2rem;
   }
 
   &__meta {
+    width: 100%;
+    margin-top: 2rem;
+
     th,
     td {
-      padding: 0.5rem 0;
+      padding: 1rem;
+      border-top: var(--color-charcoal) 1px solid;
     }
 
     th {
       vertical-align: top;
       padding-right: 1rem;
+      // font-weight: normal;
+      // color: var(--color-charcoal);
+    }
+
+    td {
+      width: 100%;
     }
   }
 
@@ -31,9 +43,10 @@
 
   &__script {
     &-content {
-      background: silver;
-      padding: 0.5rem;
-      margin: 1rem -0.5rem 0;
+      background: var(--color-platinum);
+      border-radius: 1rem;
+      padding: 1rem;
+      margin: 1rem -1rem 0;
       min-height: 10rem;
     }
   }
