+++ src/styles/_reset.scss
@@ -32,12 +32,14 @@ body {
   overflow: visible;
   height: auto;
   width: auto;
+  background: white;
 
   body,
   #root {
     overflow: visible;
     height: auto;
     width: auto;
+    background: white;
   }
 
   #preload {
